Move repositories from .Data-project to .Infrastructure-project
